usb: phy-msm-usb: Improve the PHY suspend sequence
The controller aborts the PHY suspend during device bus suspend if
there is an incoming reset/resume from the host. Abort the LPM
sequence and enable the interrupt to handle the reset/resume.
The current code resets the USB hardware and it would result in
device disconnection in this scenario.

It is observed that PHCD bit is not getting set with in the specified
time during multiple plug in and plug out tests. If the LPM is not
initiated in the device bus suspend, retry setting the PHCD bit again.
The results indicate that the PHY suspend works after the 1st attempt.
If this fails for 3 times, reset the USB hardware and retry the LPM
sequence.

The PHY suspend timeout value is modified from 500 msec to 100 usec
for device bus suspend and 5 msec for other other scenarios as per the
databook recommendation.

CRs-Fixed: 747392
Change-Id: Ie76d7a0eaff2fe40357e3af92014aa0b3b17a88c
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Ravi Vembu <raviv@motorola.com>
Reviewed-on: http://gerrit.mot.com/717215
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
(cherry picked from commit 02e74262ec5133a4c6e6c1997b6909907ae8a402)