prima: TDLS peer can get sta index as 3
Sta indexes for different interfaces and peers are as follows.
0 is for broadcast sta
1 is for self sta of wlan0
2 is for self sta of p2p0

TDLS peer can get sta index as 3 in below scenario.

p2p interface connects first and sta index 3 will be assigned to p2p
peer. If connection is initiated on wlan0 interface, wlan peer will
get sta index as 4. After which initiate the disconnection on p2p
interface, this will remove the peer with sta index 3.
At the same time if we start TDLS connection, the TDLS peer will
get the sta index as 3 and 3 is not considered as a valid TDLS peer
sta index.

To mitigate the same consider 3 as valid TDLS peer sta index.

Change-Id: I62fdbe0166c541ec88a1f276e8bc25eddde9660c
CRs-Fixed: 840109